Update spec to final VEX-WG document
This PR updates the openvex spec to match the final minimal
requirements for VEX as defined by the VEX Working group and
published by CISA on https://www.cisa.gov/sites/default/files/2023-04/minimum-requirements-for-vex-508c.pdf

The summary of changes to the spec are as follows. Except for
the document `version` changing from string to integere, all
fields are new so should not break existing consumers.

version is now an integer
added last_updated for changes to record changes to the doc
author is now recommended to be a machine readable identifier
role is now optional :tada:

@id takes an IRI to identify the statement and make it referenceable
version integer field, accounts for changes in the statement, defaults to zero and may be ommitted when so.
last_updated added a last updated timestamp to the statement
supplier an identifier of the supplier of the document

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>